Reworked package structure
