+++ Flask_Blog/flaskblog/__init__.py
@@ -1,0 +1,9 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy 
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = 'a6babd75a1da05b8e2b4ec9c0ed1c9fb'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
+db = SQLAlchemy(app)
+
+from flaskblog import routes
