+++ Flask_Blog/flaskblog/templates/home.html
